backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
django-cors-headers==3.7.0
django==3.1.5
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.2
pyjwt==2.0.1; python_version >= '3.6'
pytz==2020.5
sqlparse==0.4.1; python_version >= '3.5'
